+++ src/types/EventTarget.js
@@ -1,5 +1,4 @@
 import { dispatch, off, on } from '../modules/core/event.js'
-import { registerMethods } from '../utils/methods.js'
 import Base from './Base.js'
 
 export default class EventTarget extends Base {
@@ -57,34 +56,3 @@ export default class EventTarget extends Base {
 
   removeEventListener () {}
 }
-
-// Add events to elements
-const methods = [ 'click',
-  'dblclick',
-  'mousedown',
-  'mouseup',
-  'mouseover',
-  'mouseout',
-  'mousemove',
-  'mouseenter',
-  'mouseleave',
-  'touchstart',
-  'touchmove',
-  'touchleave',
-  'touchend',
-  'touchcancel' ].reduce(function (last, event) {
-  // add event to Element
-  const fn = function (f) {
-    if (f === null) {
-      off(this, event)
-    } else {
-      on(this, event, f)
-    }
-    return this
-  }
-
-  last[event] = fn
-  return last
-}, {})
-
-registerMethods('Element', methods)